рефакторинг в articleForm
